Add files via upload
A lot of Different Analyses are done up to "Flat-Dist, 0.1-2.0 GeV, Oct 1." Starting from Oct 4 most of these are dropped for readability and due to the prioritization of dE/dx PID of protons and pions. There is a widget that plots all the attributes of the root file- the bins and labels of all the histograms are detailed in the code. In order for it to work, "%matplotlib widget" must be placed in the cell which contains the widget code (or just so it takes precedence over "%matplotlib inline").